+++ example/lib/benchmarks/context_watch/common/observer.dart
@@ -1,6 +1,3 @@
-// ignore_for_file: unused_element
-
-// import 'package:context_plus/context_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provide_it/provide_it.dart';
@@ -68,7 +65,6 @@ class Observer extends StatelessWidget {
 
 class _ContextWatchValueListenable extends StatelessWidget {
   const _ContextWatchValueListenable({
-    super.key,
     required this.valueListenables,
     required this.visualize,
   });
@@ -100,7 +96,6 @@ class _ContextWatchValueListenable extends StatelessWidget {
 
 class _ContextWatchStream extends StatelessWidget {
   const _ContextWatchStream({
-    super.key,
     required this.streams,
     required this.visualize,
   });
@@ -137,7 +132,6 @@ class _ContextWatchStream extends StatelessWidget {
 
 class _ListenableBuilder extends StatelessWidget {
   const _ListenableBuilder({
-    super.key,
     required this.listenables,
     required this.visualize,
   });
@@ -176,7 +170,6 @@ class _ListenableBuilder extends StatelessWidget {
 
 class _ValueListenableBuilder extends StatelessWidget {
   const _ValueListenableBuilder({
-    super.key,
     required this.valueListenables,
     required this.visualize,
   });
@@ -216,7 +209,6 @@ class _ValueListenableBuilder extends StatelessWidget {
 
 class _StreamBuilder extends StatelessWidget {
   const _StreamBuilder({
-    super.key,
     required this.streams,
     required this.visualize,
   });