+++ example/lib/main.dart
@@ -1,106 +1,71 @@
 import 'package:flutter/material.dart';
 import 'package:provide_it/provide_it.dart';
-import 'package:provide_it_example/benchmarks/context_watch/benchmark_screen.dart';
+
+final pathParameters = <String, String>{
+  'counterId': 'my-counter-id',
+};
 
 void main() {
-  // readIt.provide(create);
-  readIt.provide<Abstract>(AbstractImpl.new);
-
-  final abstract = readIt.read<Abstract>();
-
   runApp(
     ProvideIt(
-      child: Builder(
-        builder: (context) {
-          final abstract = context.read<Abstract>();
-          return MaterialApp(
-            home: Scaffold(
-              body: Center(
-                child: BenchmarkScreen(),
+      // Auto-injects dependencies
+      provide: (context) {
+        context.provide(CounterService.init); // <- async
+        context.provide(CounterRepository.new);
+        context.provide(Counter.new);
+      },
+      // Auto-injects path parameters
+      namedLocator: (param) => pathParameters[param.name],
+
+      // ProvideIt will take care of loading/error, but you can customize it:
+      // - loadingBuilder: (context) => (...),
+      // - errorBuilder: (context, error, stackTrae) =>ce(...),
+      child: MaterialApp(
+        home: Builder(
+          builder: (context) {
+            final counter = context.watch<Counter>();
+
+            context.listen<Counter>((counter) {
+              // do something
+            });
+
+            return Center(
+              child: ElevatedButton(
+                onPressed: counter.increment,
+                child: Text('${counter.count}'),
               ),
-            ),
-          );
-        },
+            );
+          },
+        ),
       ),
     ),
   );
 }
 
-abstract class Abstract {}
+class CounterService {
+  CounterService();
 
-class AbstractImpl extends Abstract {}
-
-class ServiceA {
-  static Future<ServiceA> init() async {
-    await Future.delayed(Duration(seconds: 1));
-    return ServiceA();
+  static Future<CounterService> init() async {
+    await Future.delayed(Duration(seconds: 3));
+    return CounterService();
   }
 }
 
-class ServiceB {}
-
-class ServiceC {}
-
-class RepositoryA {
-  RepositoryA(this.a, this.b, this.c);
-
-  final ServiceA a;
-  final ServiceB b;
-  final ServiceC c;
+class CounterRepository {
+  CounterRepository(this.service);
+  final CounterService service;
 }
 
-class RepositoryB {
-  RepositoryB(this.a, this.b);
+class Counter extends ChangeNotifier {
+  Counter(this.repository, {required this.counterId});
+  final CounterRepository repository;
+  final String counterId;
 
-  final ServiceA a;
-  final ServiceB b;
-}
+  int _count = 0;
+  int get count => _count;
 
-class StoreA {
-  StoreA(this.a, this.b);
-
-  final RepositoryA a;
-  final RepositoryB b;
-}
-
-class StoreB {
-  StoreB(this.a);
-
-  final RepositoryA a;
-}
-
-final countRef = ValueRef<int>(0);
-
-class CounterValue extends StatelessWidget {
-  const CounterValue({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final (count, setValue) = context.value(0);
-    final (count2, setValue2) = context.value(0);
-    final (count3, setValue3) = context.value(0);
-
-    return Column(
-      children: [
-        ElevatedButton(
-          onPressed: () {
-            setValue(count + 1);
-          },
-          child: Text('Counter: $count'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            setValue2(count2 + 1);
-          },
-          child: Text('Counter2: $count2'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            setValue3(count3 + 1);
-          },
-          child: Text('Counter3: $count3'),
-        ),
-      ],
-    );
+  void increment() {
+    _count++;
+    notifyListeners();
   }
 }